refactor: fold SensorTermBase into common.manager_terms (#1371)
common/sensor_terms.py held only the binding base class and no actual terms,
so the module name and contents disagreed. SensorTermBase is generic
manager-term infrastructure (its consumers include a termination term), so it
moves to common/manager_terms.py next to the other shared term bases and
plain state/action terms; sensor_reward_terms.py keeps only the sensor reward
equations. g1/t800/sensor_reward_terms imports follow.
